requirements.txt
Flask
requests
pymongo
redis
python-dotenv
